refactor(torii): torii-runner
