+++ Cargo.toml
@@ -125,6 +125,7 @@ torii-graphql = { path = "crates/torii/graphql" }
 torii-grpc = { path = "crates/torii/grpc" }
 torii-relay = { path = "crates/torii/libp2p" }
 torii-server = { path = "crates/torii/server" }
+torii-runner = { path = "crates/torii/runner" }
 
 # saya
 saya-core = { path = "crates/saya/core" }
