+++ bin/torii/Cargo.toml
@@ -3,60 +3,9 @@ edition.workspace = true
 name = "torii"
 version.workspace = true
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-anyhow.workspace = true
-async-trait.workspace = true
-base64.workspace = true
-camino.workspace = true
-chrono.workspace = true
-clap.workspace = true
-clap_complete.workspace = true
-ctrlc = { version = "3.4", features = [ "termination" ] }
-dojo-metrics.workspace = true
-dojo-types.workspace = true
-dojo-utils.workspace = true
-dojo-world.workspace = true
-either = "1.9.0"
-futures.workspace = true
-http-body = "0.4.5"
-http.workspace = true
-hyper-reverse-proxy = { git = "https://github.com/tarrencev/hyper-reverse-proxy" }
-hyper.workspace = true
-indexmap.workspace = true
-lazy_static.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-sqlx.workspace = true
-starknet-crypto.workspace = true
-starknet.workspace = true
-tokio-stream = "0.1.11"
-tokio-util = "0.7.7"
-tokio.workspace = true
-toml.workspace = true
 torii-cli.workspace = true
-torii-indexer.workspace = true
-torii-sqlite.workspace = true
-torii-graphql.workspace = true
-torii-grpc = { workspace = true, features = [ "server" ] }
-torii-relay.workspace = true
-torii-server.workspace = true
-tower.workspace = true
-
-clap_config = "0.1.1"
-tempfile.workspace = true
-tower-http.workspace = true
-tracing-subscriber.workspace = true
-tracing.workspace = true
-url.workspace = true
-webbrowser = "0.8"
-
-[dev-dependencies]
-assert_matches.workspace = true
-camino.workspace = true
-
-[features]
-default = [ "jemalloc", "sqlite" ]
-jemalloc = [ "dojo-metrics/jemalloc" ]
-sqlite = [ "sqlx/sqlite" ]
+torii-runner.workspace = true
+tokio.workspace = true
+anyhow.workspace = true
+clap.workspace = true
